Move form-core tests
