+++ packages/form-core/tests/formOptions.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { FormApi, formOptions } from '../index'
+import { FormApi, formOptions } from '../src/index'
 
 describe('formOptions', () => {
   it('should allow default values to be set', async () => {
