+++ packages/form-core/tsconfig.docs.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["src/tests/**/*"]
+  "exclude": ["tests"]
 }
