+++ packages/form-core/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "moduleResolution": "Bundler"
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx", ".eslintrc.cjs", "vite.config.ts"]
+  "include": ["src", "tests", ".eslintrc.cjs", "vite.config.ts"]
 }
