+++ packages/form-core/tsconfig.legacy.json
@@ -3,6 +3,5 @@
   "compilerOptions": {
     "moduleResolution": "Node"
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["src/tests/**"]
+  "include": ["src"]
 }
